Make unroll array false by default
