+++ tests/tests/_folders.rs
@@ -3,4 +3,3 @@ mod cpython_benchmarks;
 mod cpython_reference_project;
 mod csharp_benchmarks;
 mod csharp_reference_project;
-mod csharp_reference_project_arrays;
