+++ tests/tests/csharp_reference_project/Test.CharArray.cs
